pyproject.toml
# This file is a part of the source code for eoc.core.
# This project has been licensed under the MIT license.
# Copyright (c) 2025-present pygame-community


[project]
name = "eoc.core"
version = "0.1.0"
description = "A library setting the foundation for EOC (Era of Challenges) infrastructure, based on the `warelib` library."
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "warelib @ git+https://github.com/pygame-community/warelib.git",
    "pygame-ce>=2.5.2",
    "pygame-gui>=0.6.13",
]

[dependency-groups]
test = ["pytest~=8.3.4", "pytest-cov~=6.0.0"]
lint = ["ruff~=0.14.10", "pre-commit~=4.0"]
dev = [
    { include-group = "test" },
    { include-group = "lint" },
    "poethepoet~=0.32",
]

[tool.poe]
executor.type = "uv"

[tool.poe.tasks]
test = "pytest src --doctest-modules --cov=src"
fmt = "ruff format ."
fmt-check = "ruff format --check ."
isort = "ruff check --select I --fix ."
lint = "ruff check ."
lint-fix = "ruff check --fix ."
formatlint = ["fmt", "isort", "lint"]
check = ["fmt-check", "lint", "test"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/eoc"]

[tool.hatch.metadata]
allow-direct-references = true

# Use Ruff to replace Black, isort, and Flake8
[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
]
ignore = [
    "E501", # line too long, handled by ruff format
    "B008", # do not perform function calls in argument defaults
    "B028", # No explicit `stacklevel` keyword argument found (?)
    "C901", # too complex
]

unfixable = ["B"] # Don't bother trying to fix flake8-bugbear

# Ignore `E402` (import violations) in all `__init__.py` files, and in select subdirectories.
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["E402"]
